Add "alphabetized" program
Preview the document and construct a Java program that read the file. After reading and storing the file, print out all of the words in alphabetical order and how many times they are used. Then print out all the words in order by the number of times they are used.