Add redblacktree java program
Red black tree: Include code necessary for timing of the load of the tree, the time to add nodes, and the time to delete nodes, from taking input from 3 txt files. Provide the code and a report for the timings for the three functions for each of the two trees.