Create Bag.java
Write a program that will loads a graph with the contents of the file Routes.csv (airline code, from airport, and to airport).  

Accept input from the console that will consist of two airport codes, determine only if it is possible to get from the one airport to the other, generate a response that is either that it is or is not possible.